pyproject.toml
[project]
name = "scooters_data_generator"
version = "1.0.0"
description = "A data generator for scooter analytics"
authors = [
    {name = "inzhenerka.tech", email = "info@inzhenerka.tech"}
]
dependencies = [
    "numpy==1.26.4",
    "osmnx==1.9.1",
    "networkx==3.2.1",
    "folium==0.16.0",
    "simpy==4.1.1",
    "faker==24.3.0",
    "pydantic==2.6.4",
    "shapely==2.0.3",
    "pandas==2.2.1",
    "geopandas==0.14.3",
    "matplotlib==3.8.3",
    "awswrangler==3.7.2",
    "pg8000==1.30.5",
    "python-dotenv==1.0.1",
    "phonenumbers==8.13.33"
]
requires-python = ">=3.11"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
